at: Use git to enhance package version
Whilst developing after a production or rc tag it would be useful to
have the package versioned with more granularity than just the last tag
or last committed version number. So use git describe subject to the
following logic:-
If git is not available, is not initialised or no commit has been made,
then the local variable pkg_version will be used. This must always be
set.
If git is available but no Tag has been created then just a commit
identifier will be used.
If git is available and a tag has been created then the most recent tag,
plus number of commits since the tag, plus a commit identifier will be
used.

Signed-off-by: Mark Grant <m.grant.prg@gmail.com>